Changed as per review comments
