+++ src/app/directives/sort-table-column-directive.ts
@@ -1,0 +1,24 @@
+import { Directive, Input, Output, EventEmitter } from "@angular/core";
+import { SortColumn, SortDirection, SortEvent } from "../users/users.component";
+
+@Directive({
+    selector: 'th[sortable]',
+    host: {
+        '[class.asc]': 'direction === "asc"',
+        '[class.desc]': 'direction === "desc"',
+        '(click)': 'rotate()'
+    }
+})
+export class NgbdSortableHeader {
+
+    @Input() sortable: SortColumn = '';
+    @Input() direction: SortDirection = '';
+    @Output() sort = new EventEmitter<SortEvent>();
+
+    rotate() {
+        this.direction = rotate[this.direction];
+        this.sort.emit({ column: this.sortable, direction: this.direction });
+    }
+}
+
+const rotate: { [key: string]: SortDirection } = { 'asc': 'desc', 'desc': '', '': 'asc' };
