+++ src/app/users/users.component.ts
@@ -7,12 +7,12 @@ import { UsersService } from '../services/users.service';
 import { IUserItem } from './user-item-model';
 import { MdbModalRef, MdbModalService } from 'mdb-angular-ui-kit/modal';
 import { CreateUserComponent } from './create-user/create-user.component';
-
+import { NgbdSortableHeader } from '../directives/sort-table-column-directive';
 
 
 export type SortColumn = keyof IUserItem | '';
 export type SortDirection = 'asc' | 'desc' | '';
-const rotate: { [key: string]: SortDirection } = { 'asc': 'desc', 'desc': '', '': 'asc' };
+
 
 const compare = (v1: any, v2: any) => v1 < v2 ? -1 : v1 > v2 ? 1 : 0;
 
@@ -21,25 +21,6 @@ export interface SortEvent {
   direction: SortDirection;
 }
 
-@Directive({
-  selector: 'th[sortable]',
-  host: {
-    '[class.asc]': 'direction === "asc"',
-    '[class.desc]': 'direction === "desc"',
-    '(click)': 'rotate()'
-  }
-})
-export class NgbdSortableHeader {
-
-  @Input() sortable: SortColumn = '';
-  @Input() direction: SortDirection = '';
-  @Output() sort = new EventEmitter<SortEvent>();
-
-  rotate() {
-    this.direction = rotate[this.direction];
-    this.sort.emit({ column: this.sortable, direction: this.direction });
-  }
-}
 
 @Component({
   selector: 'app-users',
